Fix for realpath returning NULL

+++ build-all.sh
@@ -33,34 +33,37 @@ full_build_for_arch() {
 }
 
 # Clean everything.
+ARM_BUILD=1
+if [ "$ARM_BUILD" = "1" ] ; then
+    # Make arm build.
+    full_build_for_arch $PREFIX arm
+    # Remove Apple's proprietary stuff, backing up the list of what's needed.
+    find /tmp2/${PREFIX}-ios/usr/lib > /tmp2/${PREFIX}-ios/needed-libs.txt
+    find /tmp2/${PREFIX}-ios/arm-apple-darwin11/lib >> /tmp2/${PREFIX}-ios/needed-libs.txt
+    rm -rf /tmp2/${PREFIX}-ios/usr/lib
+    rm -rf /tmp2/${PREFIX}-ios/arm-apple-darwin11/lib
+    rm /tmp2/${PREFIX}-ios/lib/libSystem.B.dylib
+    find /tmp2/${PREFIX}-ios/usr/include > /tmp2/${PREFIX}-ios/needed-headers.txt
+    find /tmp2/${PREFIX}-ios/arm-apple-darwin11/sys-include >> /tmp2/${PREFIX}-ios/needed-headers.txt
+    rm -rf /tmp2/${PREFIX}-ios/usr/include
+    rm -rf /tmp2/${PREFIX}-ios/arm-apple-darwin11/sys-include
+fi
 
-# Make arm build.
-full_build_for_arch $PREFIX arm
-
-# Remove Apple's proprietary stuff, backing up the list of what's needed.
-find /tmp2/${PREFIX}-ios/usr/lib > /tmp2/${PREFIX}-ios/needed-libs.txt
-find /tmp2/${PREFIX}-ios/arm-apple-darwin11/lib >> /tmp2/${PREFIX}-ios/needed-libs.txt
-rm -rf /tmp2/${PREFIX}-ios/usr/lib
-rm -rf /tmp2/${PREFIX}-ios/arm-apple-darwin11/lib
-rm /tmp2/${PREFIX}-ios/lib/libSystem.B.dylib
-find /tmp2/${PREFIX}-ios/usr/include > /tmp2/${PREFIX}-ios/needed-headers.txt
-find /tmp2/${PREFIX}-ios/arm-apple-darwin11/sys-include >> /tmp2/${PREFIX}-ios/needed-headers.txt
-rm -rf /tmp2/${PREFIX}-ios/usr/include
-rm -rf /tmp2/${PREFIX}-ios/arm-apple-darwin11/sys-include
-
-# Make i686 build.
-full_build_for_arch $PREFIX intel
-
-# Remove Apple's proprietary stuff, backing up the list of what's needed.
-find /tmp2/${PREFIX}-osx/usr/lib > /tmp2/${PREFIX}-osx/needed-libs.txt
-find /tmp2/${PREFIX}-osx/i686-apple-darwin11/lib >> /tmp2/${PREFIX}-osx/needed-libs.txt
-rm -rf /tmp2/${PREFIX}-osx/usr/lib
-rm -rf /tmp2/${PREFIX}-osx/i686-apple-darwin11/lib
-rm /tmp2/${PREFIX}-osx/lib/libSystem.B.dylib
-find /tmp2/${PREFIX}-osx/usr/include > /tmp2/${PREFIX}-osx/needed-headers.txt
-find /tmp2/${PREFIX}-osx/i686-apple-darwin11/sys-include >> /tmp2/${PREFIX}-osx/needed-headers.txt
-rm -rf /tmp2/${PREFIX}-osx/usr/include
-rm -rf /tmp2/${PREFIX}-osx/i686-apple-darwin11/sys-include
+INTEL_BUILD=1
+if [ "$INTEL_BUILD" = "1" ] ; then
+    # Make i686 build.
+    full_build_for_arch $PREFIX intel
+    # Remove Apple's proprietary stuff, backing up the list of what's needed.
+    find /tmp2/${PREFIX}-osx/usr/lib > /tmp2/${PREFIX}-osx/needed-libs.txt
+    find /tmp2/${PREFIX}-osx/i686-apple-darwin11/lib >> /tmp2/${PREFIX}-osx/needed-libs.txt
+    rm -rf /tmp2/${PREFIX}-osx/usr/lib
+    rm -rf /tmp2/${PREFIX}-osx/i686-apple-darwin11/lib
+    rm /tmp2/${PREFIX}-osx/lib/libSystem.B.dylib
+    find /tmp2/${PREFIX}-osx/usr/include > /tmp2/${PREFIX}-osx/needed-headers.txt
+    find /tmp2/${PREFIX}-osx/i686-apple-darwin11/sys-include >> /tmp2/${PREFIX}-osx/needed-headers.txt
+    rm -rf /tmp2/${PREFIX}-osx/usr/include
+    rm -rf /tmp2/${PREFIX}-osx/i686-apple-darwin11/sys-include
+fi
 
 # Strip executables.
 # Maybe "strip -u -r -S" when on OS X?
